\name{merge}
\alias{merge}
\alias{merge-methods}
\alias{merge,OptimizedHierarchy,OptimizedHierarchy-method}
\title{
  Merges two OptimizedHierarchy objects.
}
\description{
  Merges two OptimizedHierarchy objects.
}
\usage{
\S4method{merge}{OptimizedHierarchy,OptimizedHierarchy}(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    First OptimizedHierarchy.
  }
  \item{y}{
    Second OptimizedHierarchy.
  }
}
\value{
  \item{OptH}{An OptimizedHierarchy object}
}


\author{
  Nima Aghaeepour \email{<naghaeep@gmail.com>} and Adrin Jalali \email{<ajalali@bccrc.ca>}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ FlowCytData }
