\name{plot}
\docType{methods}
\alias{plot}
\alias{plot,OptimizedHierarchy,ANY-method}
\title{Plots an optimized cellular hierarchy}
\description{
  Plots an optimized cellular hierarchy
}

\usage{
\S4method{plot}{OptimizedHierarchy,ANY}(x, phenotypeScores, uniformColors = FALSE, ylab = NULL, xlab = NULL, colors = c("blue", "cyan", "yellow", "red"), edgeWeights = TRUE, edgeLabels = TRUE, nodeLabels = TRUE, min.score = NA, max.score = NA, cell.proportions = NULL, min.proportion = NA, max.proportion = NA, proportion.colors = c("black", "white"), node.lwd = 5, root.name = "All Cells")
}

\arguments{
\item{x}{An OptimizedHierarchy object as generated by RchyOptimyx.}
\item{phenotypeScores}{The score vector used for determining the colors
  of the nodes.}
\item{uniformColors}{Boolean variable. If TRUE, the graph will not have
  any colors.}
\item{ylab}{y-axis label of the density plot on the right.}
\item{xlab}{x-axis label of the density plot of the phenotype cell proportion in the buttom.}
\item{colors}{Color vector indicating colors to be used in the right pallet.}
\item{edgeWeights}{Boolean value indicating weather edgeWeights should be plotted or not.}
\item{edgeLabels}{Boolean value indicating weather edgeLabels should be plotted or not.}
\item{nodeLabels}{Boolean value indicating weather nodeLabels should be plotted or not.}
\item{min.score}{Double value indicating the minimum value of scores. Default value will use minimum node scores of the given \code{OptimizedHierarchy} object.}
\item{max.score}{Double value indicating the maximum value of scores. Default value will use maximum node scores of the given \code{OptimizedHierarchy} object.}
\item{cell.proportions}{The score vector used to determine the color of node borders to illustrate phenotype cell proportions.}
\item{min.proportion}{Double value indicating the minimum value of the cell proportions. Default value will use minimum cell proportion of the given \code{OptimizedHierarchy} object.}
\item{max.proportion}{Double value indicating the minimum value of the cell proportions. Default value will use minimum cell proportion of the given \code{OptimizedHierarchy} object.}
\item{proportion.colors}{Color vector indicating colors to be used in the buttom pallet.}
\item{node.lwd}{Integer value indicating node border width.}
\item{root.name}{Character value indicating the root node's name.}
}

\author{
  Nima Aghaeepour \email{<naghaeep@bccrc.ca>} and Adrin Jalali \email{<ajalali@bccrc.ca>}
}

\seealso{
\code{\link{RchyOptimyx-methods}}
}

\keyword{print}
