\name{plot}
\docType{methods}
\alias{plot}
\alias{plot,OptimizedHierarchy,ANY-method}
\alias{plot,OptimizedHierarchy-method}
\title{Plots an optimized cellular hierarchy}
\description{
  Plots an optimized cellular hierarchy
}

\usage{
\S4method{plot}{OptimizedHierarchy,ANY}(x, phenotypeScores, phenotypeCodes, marker.names,
                 partitions.per.marker = NULL, 
uniformColors = FALSE, ylab = NULL, xlab = NULL, colors = c("blue",
"cyan", "yellow", "red"), edgeWeights = TRUE, edgeLabels = TRUE,
nodeLabels = TRUE, min.score = NA, max.score = NA, cell.proportions =
NULL, min.proportion = NA, max.proportion = NA, proportion.colors =
c("black", "white"), node.lwd = 5, root.name = "All Cells",
legend.size=1.25, plot.legend=TRUE, textcol=par('fg'))
}

\arguments{
\item{x}{An OptimizedHierarchy object as generated by RchyOptimyx.}
\item{phenotypeScores}{The score vector used for determining the colors
  of the nodes.}
\item{phenotypeCodes}{A vector of strings of length N (the number of markers) for each phenotype measured. For every phenotype, the character corresponding to a given marker can be 0, 1, 2, etc for
    neutral, negative, positive, bright, etc. See
    the provided vigenette for more details
    and examples.}
\item{marker.names}{A vector of characters representing the names of all markers involved in the analysis.}
\item{partitions.per.marker}{A vector of integers, one per marker, describing the number of partitions per marker. Default is 2 (positive and negative) for all markers.}
\item{uniformColors}{Boolean variable. If TRUE, the graph will not have
  any colors.}
\item{ylab}{y-axis label of the density plot on the right.}
\item{xlab}{x-axis label of the density plot of the phenotype cell proportion in the buttom.}
\item{colors}{Color vector indicating colors to be used in the right pallet.}
\item{edgeWeights}{Boolean value indicating weather edgeWeights should be plotted or not.}
\item{edgeLabels}{Boolean value indicating weather edgeLabels should be plotted or not.}
\item{nodeLabels}{Boolean value indicating weather nodeLabels should be plotted or not.}
\item{min.score}{Double value indicating the minimum value of scores. Default value will use minimum node scores of the given \code{OptimizedHierarchy} object.}
\item{max.score}{Double value indicating the maximum value of scores. Default value will use maximum node scores of the given \code{OptimizedHierarchy} object.}
\item{cell.proportions}{The score vector used to determine the color of node borders to illustrate phenotype cell proportions.}
\item{min.proportion}{Double value indicating the minimum value of the cell proportions. Default value will use minimum cell proportion of the given \code{OptimizedHierarchy} object.}
\item{max.proportion}{Double value indicating the minimum value of the cell proportions. Default value will use minimum cell proportion of the given \code{OptimizedHierarchy} object.}
\item{proportion.colors}{Color vector indicating colors to be used in the buttom pallet.}
\item{node.lwd}{Integer value indicating node border width.}
\item{root.name}{Character value indicating the root node's name.}
\item{legend.size}{Size of the color legend in inches.}
\item{plot.legend}{Boolean indicating wether the color legend should be plotted.}
\item{textcol}{The color of the axis label(s) for the color legend(s).}
}

\author{
  Nima Aghaeepour \email{<naghaeep@stanford.edu>} and Adrin Jalali \email{<ajalali@bccrc.ca>}
}

\seealso{
\code{\link{RchyOptimyx-methods}}
}

\keyword{print}
